+++ src/file-formats/common/parse-utils.ts
@@ -1,7 +1,10 @@
-import { ReadTFileArgs } from "../file-format-definitions";
+import { ReadTFileArgs, WriteTFileArgs } from "../file-format-definitions";
 import { getDebugPath, logFatal } from "../../util/util";
 
-export function logParseError(rawMsg: string, args: ReadTFileArgs): never {
+export function logParseError(
+  rawMsg: string,
+  args: ReadTFileArgs | WriteTFileArgs
+): never {
   const msg = `Failed to parse ${getDebugPath(
     args.path
   )} with expected format '${args.format}': ${rawMsg}`;